local_server/requirements.txt
# requirements.txt
fastapi
uvicorn
python-multipart
nemo_toolkit[asr]
torch
torchaudio
pydub
